Reduce codes in App and move them to Backend
