--- src/Files.App/CommandLine/ParsedCommands.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace Files.App.CommandLine
-{
-	internal class ParsedCommands : List<ParsedCommand>
-	{
-	}
-}
