+++ src/Files.App/Converters/EnumConverters/ConflictResolveOptionToIndexConverter.cs
