+++ src/Files.App/Converters/GenericEnumConverter.cs
