+++ src/Files.App/Converters/ImageModelToImageConverter.cs
@@ -1,4 +1,4 @@
-using Files.App.AppModels;
+using Files.App.DataModels;
 using Microsoft.UI.Xaml.Data;
 using System;
