+++ src/Files.App/Converters/StringToBrushConverter.cs
