+++ src/Files.App/DataModels/BitmapImageModel.cs
@@ -1,7 +1,7 @@
 ﻿using Files.Backend.Models;
 using Microsoft.UI.Xaml.Media.Imaging;
 
-namespace Files.App.AppModels
+namespace Files.App.DataModels
 {
 	/// <inheritdoc cref="IImageModel"/>
 	internal sealed class BitmapImageModel : IImageModel
