+++ src/Files.App/Helpers/ThemeHelper.cs
@@ -1,5 +1,5 @@
 using CommunityToolkit.WinUI;
-using Files.App.Extensions;
+using Files.Backend.Extensions;
 using Microsoft.UI;
 using Microsoft.UI.Windowing;
 using Microsoft.UI.Xaml;
