+++ src/Files.App/Program.cs
@@ -1,6 +1,6 @@
-using Files.App.CommandLine;
 using Files.App.Helpers;
 using Files.App.Shell;
+using Files.Backend.CommandLine;
 using Files.Backend.Helpers;
 using Files.Shared.Extensions;
 using Microsoft.UI.Dispatching;
