+++ src/Files.Backend/CommandLine/ParsedCommands.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace Files.Backend.CommandLine
+{
+	public class ParsedCommands : List<ParsedCommand>
+	{
+	}
+}
