+++ src/Files.Backend/Helpers/AsyncManualResetEvent.cs
