+++ src/Files.Backend/Helpers/CollectionDebugView.cs
