+++ src/Files.Backend/Helpers/EnumConversionHelpers.cs
