+++ src/Files.Backend/Helpers/IntervalSampler.cs
