+++ src/Files.Backend/Serialization/BaseJsonSettings.cs
