+++ src/Files.Backend/Serialization/BaseObservableJsonSettings.cs
