+++ src/Files.Backend/Serialization/IJsonSettingsDatabase.cs
