+++ src/Files.Backend/Serialization/ISettingsSharingContext.cs
